Extract SimpleClient::Management to its own file
